requirements.txt
bleak
paho-mqtt
requests